전투 데이터 로딩 구조 개편
- 전투 시작 로직(CombatService)을 UserCharacterId 기반에서 CharacterId 기반으로 전면 수정
- UserCharacters 테이블 구조에 맞게 스탯 조회 로직을 CharacterId 기준으로 정리
- EfUserCharacterReader에 GetManyByCharacterIdAsync 추가 및 실제 사용하도록 전체 흐름 수정
- BuildPackAsync가 UserCharacterId를 사용하는 문제 해결: CharacterId + UserId 기반으로 스탯 로딩하도록 변경
- CombatMasterDataPack 생성 시 ActorDef 키를 CharacterId 기준으로 통일
- 파티 슬롯(UserPartyCharacter)은 기존처럼 characterId를 사용하되, 스텟/패키지만 CharacterId 기반으로 정합성 확보
- KeyNotFoundException(‘15’) 오류 해결: 모든 조회 흐름을 CharacterId 단일 기준으로 통일함

+++ WebServer/Application/Combat/CombatMasterModels.cs
@@ -87,11 +87,11 @@ namespace Application.Combat
     public sealed class CombatMasterDataPack
     {
         public CombatStageDef Stage { get; }
-        public IReadOnlyDictionary<int, CombatActorDef> Actors { get; }
+        public IReadOnlyDictionary<long, CombatActorDef> Actors { get; }
 
         public CombatMasterDataPack(
             CombatStageDef stage,
-            IReadOnlyDictionary<int, CombatActorDef> actors)
+            IReadOnlyDictionary<long, CombatActorDef> actors)
         {
             Stage = stage;
             Actors = actors;
